Changed to reflect new guidelines.

+++ TAO/examples/Simple/bank/run_test.pl
@@ -8,16 +8,27 @@ eval '(exit $?0)' && eval 'exec perl -S $0 ${1+"$@"}'
 use lib "../../../../bin";
 require ACEutils;
 require Process;
+$status = 0;
 
 $iorfile = "grid.ior";
 
 $SV = Process::Create ($EXEPREFIX."server$EXE_EXT", "-o $iorfile ");
 
-ACE::waitforfile ($iorfile);
+if (ACE::waitforfile_timed ($iorfile, 5) == -1) {
+  print STDERR "ERROR: timedout waiting for file <$iorfile>\n";
+  $SV->Kill (); $SV->TimedWait (1);
+  exit 1;
+}
 
-$status  = system ($EXEPREFIX."client$EXE_EXT -f $iorfile");
-
-$SV->Kill (); $SV->Wait ();
+$client = Process::Create($EXEPREFIX."client$EXE_EXT","-f $iorfile");
+if ($client->TimedWait (60) == -1) {
+  print STDERR "ERROR: the client timedout\n";
+  $status = 1;
+  $client->Kill (); $client->TimedWait (1);
+}
+  
+  
+$SV->Kill (); $SV->TimedWait (1);
 
 unlink $iorfile;
