+++ TAO/examples/Simple/grid/run_test.pl
@@ -9,15 +9,26 @@ use lib "../../../../bin";
 require ACEutils;
 require Process;
 
+$status = 0;
 $iorfile = "grid.ior";
 
 $SV = Process::Create ($EXEPREFIX."server$EXE_EXT", "-o $iorfile ");
 
-ACE::waitforfile ($iorfile);
+if (ACE::waitforfile_timed ($iorfile, 5) == -1) {
+  print STDERR "ERROR: timedout waiting for file <$iorfile>\n";
+  $SV->Kill (); $SV->TimedWait (1);
+  exit 1;
+}
 
-$status  = system ($EXEPREFIX."client$EXE_EXT -f $iorfile -x -w 12 -h 12 -p 2 -q 2 -v 2345");
+$client = Process::Create($EXEPREFIX."client$EXE_EXT", "-f $iorfile -x -w 12 -h 12 -p 2 -q 2 -v 2345");
 
-$SV->Kill (); $SV->Wait ();
+if ($client->TimedWait (60) == -1) {
+  print STDERR "ERROR: client timedout\n";
+  $status = 1;
+  $client->Kill (); $client->TimedWait (1);
+}
+
+$SV->Kill (); $SV->TimedWait (1);
 
 unlink $iorfile;
